Seperating the surface from the builder, in service of upcoming features.
