--- addons/beziersurfaces/Scripts/surface_script.gd.uid
@@ -1,1 +1,0 @@
-uid://bu3qs2cd5sgv4
